Changes CORS settings in app.py
Copies over 1-hbnb.py to 2-hbnb.py and creates a /2-hbnb route that renders 2-hbnb.html.

Added 2-hbnb.js that queries the flask server on PORT 5001. Adds the class "available" when the GET request returns a 200 status code.

Creates CSS for a div with id=api_status and CSS for a class called available when the GET jQuery returns a good response.

2-hbnb.html contains a new div with id = api_status in the header class.